drm/i915/display: Enable pipelock using kernel param
Pipelock is a display feature to sync vblank timing
for multiple display setup. Primary display need to
be set so that it's can use as reference for the
other display. We can use i915.pipelock_primary=x
where x is 0 = PIPE_A, 1 = PIPE_B, and etc to enable
pipelock capability.

Signed-off-by: Zawawi, Muhammad Zul Husni <muhammad.zul.husni.zawawi@intel.com>